Story #13: expose isDefined() function, split off RxJS operators
