+++ src/public-api.ts
@@ -8,5 +8,6 @@ export * from './lib/resource';
 export * from './lib/collection';
 export * from './lib/accessor';
 export * from './lib/utils';
+export * from './lib/rxjs';
 export * from './lib/hal-client.service';
 export * from './lib/hal-client.module';
